+++ config/di-web.php
@@ -2,8 +2,10 @@
 
 declare(strict_types=1);
 
+use Yiisoft\ErrorHandler\Factory\ThrowableResponseFactory;
 use Yiisoft\ErrorHandler\Renderer\HtmlRenderer;
 use Yiisoft\ErrorHandler\ThrowableRendererInterface;
+use Yiisoft\ErrorHandler\ThrowableResponseFactoryInterface;
 
 /**
  * @var array $params
@@ -11,4 +13,5 @@ use Yiisoft\ErrorHandler\ThrowableRendererInterface;
 
 return [
     ThrowableRendererInterface::class => HtmlRenderer::class,
+    ThrowableResponseFactoryInterface::class => ThrowableResponseFactory::class,
 ];